Add GitHub Actions workflow for auto-formatting README.md and implement project formatting script
