+++ projects.md
@@ -1,0 +1,97 @@
+<h1 dir="rtl">پروژه‌ها</h1>
+
+<p dir="rtl">این مخزن با هدف آرشیو پروژه‌های اعضای تهلاگ ایجاد شده است. هدف از این مجموعه، افزایش دیده‌شدن، حمایت متقابل و تسهیل همکاری در پروژه‌های یکدیگر است. انتظار می‌رود که این مخزن، تأثیر مثبتی بر فعالیت‌های اعضا در دنیای متن‌باز داشته باشد. برای مشارکت و افزودن پروژه به این لیست، اطلاعات و جزئیات پروژه خود را مطابق با نمونه‌ در فایل <a href='./projects.md'>projects.md</a> در یک PR ارسال کنید.
+</p>
+
+<h2 dir="rtl">htte</h2>
+<p dir="rtl">یه وب‌سرور http که از صفر با گولنگ توسعه داده می‌شه.</p>
+
+<a href="https://github.com/the-pesar/htte">لینک مخزن</a>
+
+<h2 dir="rtl">V-OTP</h2>
+<p dir="rtl">یک کامپوننت لایبرری برای Vue با هدف ارائه کامپوننت ورودی برای پسورد یکبار مصرف.</p>
+
+<a href="https://github.com/Saman-Safaei-Dev/v-otp">لینک مخزن</a>
+
+<h2 dir="rtl">Tasky</h2>
+<p dir="rtl">تسکی یک برنامه خط فرمان (CLI) است که به شما امکان می‌دهد به‌سادگی لیست وظایف خود را ایجاد و مدیریت کنید. این ابزار بصورت کامل با گولنگ توسعه پیدا کرده و با هدف فراهم کردن یک تجربه کاربری ساده و کارآمد برای مدیریت کارهای روزانه طراحی شده و با مجوز GPL-3.0 منتشر شده است.</p>
+
+<a href="https://github.com/shahriaarrr/tasky">لینک مخزن</a>
+
+<h2 dir="rtl">BoxMovie</h2>
+<p dir="rtl">وب اپلیکیشن SPA / PWA با هدف تماشای فیلم و سریال زیرنویس بدون سانسور و رایگان !</p>
+
+<a href="https://github.com/BoxMovie/BoxMovie.github.io">لینک مخزن</a> | <a href="https://boxmovie.github.io">لینک سایت</a>
+
+<h2 dir="rtl">Chat Guard</h2>
+<p dir="rtl">چت گارد یک افزونه مرورگره که امکان این رو به شما میده که بتونید داخل هر پیام رسانی تحت وب به صورت end to end چت کنید و از اطلاعاتتون حفاظت کنید.</p>
+
+<a href="https://github.com/PrivacyForge/ChatGuard">لینک مخزن</a>
+
+<h2 dir="rtl">Tehran Metro Data</h2>
+<p dir="rtl">این مخزن تمام داده مترو تهران را به صورت json و گراف در اختیار قرار داده که داخل اپلیکشین های مختلف مورد استفاده قرار بگیرد.
+این اطلاعات شامل روابط بین ایستگاه ها و تمام امکانات رفاهی ان ها می‌باشد.
+</p>
+
+<a href="https://github.com/mostafa-kheibary/tehran-metro-data">لینک مخزن</a>
+
+<h2 dir="rtl">Xcleaner</h2>
+<p dir="rtl">ایکس کلینر یک افزونه کروم هستش که به شما این امکان رو میده که بتونید صفحه x (twitter) خود را بر اساس محتوای نام کاربران فیلتر کنید
+</p>
+
+<a href="https://github.com/mostafa-kheibary/xCleaner">لینک مخزن</a>
+
+<h2 dir="rtl">squid_game</h2>
+<p dir="rtl">درحال ایجاد و توسعه بازی اسکوییدگیم با ریکت و نکست هستم امیدوارم به من تو این کار کمک کنین</p>
+
+<a href="https://github.com/mh-morowati/squid_game">لینک مخزن</a>
+
+<h2 dir="rtl">DeepSeek Desktop Application</h2>
+<p dir="rtl">DeepSeek یک برنامه دسکتاپ است که برای هوش مصنوعی DeepSeek توسعه داده شده است. این برنامه یک رابط کاربری ساده برای تعامل با پلتفرم مبتنی بر هوش مصنوعی DeepSeek فراهم می‌آورد. برنامه با استفاده از Electron ساخته شده و ویژگی‌هایی مانند دسترسی به DeepSeek Chat، رابط کاربری تمیز و قابل تنظیم و گزینه‌های تنظیمات مختلف را ارائه می‌دهد.</p>
+
+<a href="https://github.com/code3-dev/deepseek-desktop">لینک مخزن</a>
+
+<h2 dir="rtl">Bale Desktop Application</h2>
+<p dir="rtl">Bale شبکه اجتماعی-مالی بانک ملی ایران است که به طور همزمان امکان ارتباط و پرداخت را فراهم می‌آورد. با استفاده از باله، می‌توانید به راحتی تماس‌های صوتی و تصویری برقرار کرده و لحظات خود را با دوستانتان به اشتراک بگذارید. این برنامه با استفاده از Electron ساخته شده است.</p>
+
+<a href="https://github.com/code3-dev/bale-desktop">لینک مخزن</a>
+
+<h2 dir="rtl">SyntoraPHP</h2>
+<p dir="rtl">SyntoraPHP یک فریمورک PHP با عملکرد بالا است که به طور خاص برای ساخت API ها طراحی شده است. این فریمورک محیطی سریع، امن و آسان برای توسعه برنامه‌های API فراهم می‌آورد. با استفاده از SyntoraPHP، می‌توانید از سیستم مسیریابی قدرتمند مشابه لاراول بهره‌مند شوید و از اجزای اصلی مانند کلاس MySQL، کلاس HTTP و کلاس Cipher برای رمزگذاری و رمزگشایی داده‌ها استفاده کنید.</p>
+
+<a href="https://github.com/code3-dev/SyntoraPHP">لینک مخزن</a>
+
+<h2 dir="rtl">Native IG</h2>
+<p dir="rtl">Native IG یک برنامه مدرن برای دانلود ویدیوهای اینستاگرام است که با استفاده از Next.js ساخته شده و بدون مشکل کار می‌کند.</p>
+
+<a href="https://github.com/code3-dev/native-ig">لینک مخزن</a>
+
+<h2 dir="rtl">Client</h2>
+<p dir="rtl">Client یک کتابخانه PHP است که یک رابط ساده و کاربرپسند برای انجام درخواست‌های HTTP با استفاده از cURL فراهم می‌کند. این کتابخانه به شما امکان می‌دهد به راحتی هدرهای HTTP، بدنه درخواست، متد درخواست و سایر گزینه‌ها را تنظیم کنید.</p>
+
+<a href="https://github.com/httptools/Client">لینک مخزن</a>
+
+<h2 dir="rtl">Widgetify</h2>
+<p dir="rtl">
+ویجتی‌فای یه اپلیکیشن تحت دسکتاپ (ویندوز،لینوکس،مک) و وب (PWA) هست. که توی خودی لیستی از ویجت های کاربردی رو مثل قیمت ارزها(رمزارزها، طلا، دلار و...)، وضعیت آب و هوا با هوش مصنوعی، تقویم و ... 
+</p>
+
+<a href="https://github.com/widgetify-app/">لینک مخزن</a> | <a href="https://www.widgetify.ir/">لینک سایت</a>
+
+<h2 dir="rtl">DNS Changer</h2>
+<p dir="rtl">
+یک نرم‌افزار متن‌باز با بیش از 50K یوزر فعال است که با استفاده از آن، کاربران می‌توانند با سادگی تنظیمات DNS خود را تغییر دهند و به این ترتیب، از دسترسی به اینترنت با سرعت و کارایی بالاتری برخوردار شوند.
+ویندوز، لینوکس، مک
+</p>
+
+<a href="https://github.com/DnsChanger">لینک مخزن</a> | <a href="https://dnschanger.github.io">لینک سایت</a>
+
+<h2 dir="rtl">Carburetor</h2>
+<p dir="rtl">تجربهٔ مرور ناشناس بی‌دردسر با کاربراتور، سپر محرمانگی بین‌افزاره‌ای. ساخته شده برای گنو/لینوکس با پکپارچگی گنوم و قدرت گرفته از تور.</p>
+
+<a href="https://framagit.org/tractor/carburetor">پیوند مخزن</a> | <a href="https://flathub.org/fa/apps/io.frama.tractor.carburetor">پیوند پایگاه وب</a>
+
+<h2 dir="rtl">System Cooler Daemon</h2>
+<p dir="rtl">این برنامه برای نظارت بر دمای سیستم و تنظیم عامل خنک کننده بر اساس آن طراحی شده است.</p>
+
+<a href="https://github.com/LogicCavalier/SystemCooler.git">پیوند مخزن</a>
